requirements.txt
yfinance
alpaca-py
openai
python-dotenv
pandas
requests
lxml
pydantic

